+++ admin/src/main/java/org/yapp/user/application/UserService.java
@@ -13,9 +13,9 @@ import org.yapp.core.domain.profile.Profile;
 import org.yapp.core.domain.profile.ProfileRejectHistory;
 import org.yapp.core.domain.profile.ProfileValueTalk;
 import org.yapp.core.domain.user.User;
-import org.yapp.error.dto.ProfileErrorCode;
-import org.yapp.error.dto.UserErrorCode;
-import org.yapp.error.exception.ApplicationException;
+import org.yapp.exception.ApplicationException;
+import org.yapp.exception.error.code.ProfileErrorCode;
+import org.yapp.exception.error.code.UserErrorCode;
 import org.yapp.format.PageResponse;
 import org.yapp.profile.dao.ProfileRejectHistoryRepository;
 import org.yapp.profile.dao.ProfileValueTalkRepository;
