+++ api/src/main/java/org/yapp/domain/profile/application/ProfileValuePickService.java
@@ -16,8 +16,8 @@ import org.yapp.domain.profile.presentation.request.ProfileValuePickCreateReques
 import org.yapp.domain.profile.presentation.response.ProfileValuePickResponses;
 import org.yapp.domain.user.application.UserService;
 import org.yapp.domain.value.application.ValuePickService;
-import org.yapp.error.dto.ProfileErrorCode;
-import org.yapp.error.exception.ApplicationException;
+import org.yapp.exception.ApplicationException;
+import org.yapp.exception.error.code.ProfileErrorCode;
 
 @Service
 @RequiredArgsConstructor
