+++ api/src/main/java/org/yapp/domain/user/application/UserService.java
@@ -10,8 +10,8 @@ import org.yapp.core.domain.user.RoleStatus;
 import org.yapp.core.domain.user.User;
 import org.yapp.domain.auth.presentation.dto.response.OauthLoginResponse;
 import org.yapp.domain.user.dao.UserRepository;
-import org.yapp.error.dto.UserErrorCode;
-import org.yapp.error.exception.ApplicationException;
+import org.yapp.exception.ApplicationException;
+import org.yapp.exception.error.code.UserErrorCode;
 
 @Service
 @RequiredArgsConstructor
